+++ src/main/kotlin/cloud/glitchdev/rfu/achievement/achievements/trophy/FishPityAchievement.kt
@@ -19,7 +19,7 @@ object FishPityAchievement : BaseAchievement() {
     override val category: AchievementCategory = AchievementCategory.TROPHY_FISHING
 
     override fun setupListeners() {
-        activeListeners.add(registerTrophyFishCatchEvent(priority = 10) { fish, tier ->
+        activeListeners.add(registerTrophyFishCatchEvent(priority = 10) { fish, tier, _ ->
             if (tier == TrophyTier.DIAMOND) {
                 val existing = TrophyDataManager.data.pity.fishPity[fish.name]
                 if (existing != null && existing.diamondProgress >= fish.diamondPity) {