+++ CHANGELOG.md
@@ -1,4 +1,11 @@
 # v1.16.0 - ???
 
 ### Fixes
-- Added more validations before registering the hud to prevent a somewhat rare crash.
+- Fixed command suggestion box width calculation conflict with SkyHanni for emojis.
+- Fixed trophy fish and frog pity display counting as one when caught with a multiplier (e.g. Geometric).
+- Fixed resource pack feature not updating the resourcepacks and saving because of file locks.
+
+### Changes
+- Set the minimum required number of chars to display the emoji suggestions to 2 and made it configurable on Other Settings.
+- Added a setting to toggle whether Enter completes emoji suggestions.
+- Improved Search on the settings so it doesn't need to open the category